net: fix device conformance sections
For the network device, we had two device normative sections both called
"setting up receive buffers", neither of which was referenced in the
conformance section.

Let's rename the second one to "processing of packets" which seems to
better match the actual contents and reference both of them from the
conformance statement for network devices.

Resolves VIRTIO-97.

Agreed on the 2014/05/06 TC meeting.

Signed-off-by: Cornelia Huck <cornelia.huck@de.ibm.com>

git-svn-id: https://tools.oasis-open.org/version-control/svn/virtio@362 0c8fb4dd-22a2-4bb5-bc14-6c75a5f43652