legacy: consistently use past tense
Paragraph with general description of feature negotiation
for legacy devices mixed present and past tense.
As rest of legacy sections all use past tense,
fix the only instance of the present tense:
s/do/did/ for consistency.

It might be argued that legacy devices still have these
properties so present tense is more appropriate, on the
other hand, using the past tense helps stress the fact
that current spec does not attempt to fully describe the legacy
device/driver behaviour: this text is only here to serve as
motivation for the transitional device/driver requirements.

VIRTIO-112

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

git-svn-id: https://tools.oasis-open.org/version-control/svn/virtio@413 0c8fb4dd-22a2-4bb5-bc14-6c75a5f43652