makediff: work around for lstlisting
latexdiff seems to go wild trying to diff lstlisting.
But marking lstlisting as picture - so old text is
commented out - then removing the comment sign -
seems to work, with the only disadvantage being
that listing is diffed as a whole.
Seems to be the best I can come up with ATM.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

git-svn-id: https://tools.oasis-open.org/version-control/svn/virtio/branches/v1.0@484 0c8fb4dd-22a2-4bb5-bc14-6c75a5f43652