net: clarify device rules for mergeable buffers
The idea behind mergeable buffers was to simply use them in a way
similar to a chain of descriptors.  Unfortunately the current text does
not say so - apparently nothing says device can't spread a packet over
as many buffers as it likes - but this didn't prevent drivers from
relying on buffers being used as a chain of descriptors, completely -
and blindly accessing it without checking the length at least for the
packet header.

Let's just make the spec match this reality - if devices ever want more
flexibility, we can add a feature bit.

Further, correct all misuses of a "descriptor" to "buffer" as that
is the entity that is being merged.

VIRTIO-160

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Message-Id: <1490655576-12173-1-git-send-email-mst@redhat.com>
Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>