introduction: simplify the designation of legacy
The sentence designating the documents defining what later
became known as the legacy virtio interface had the most
important piece of information placed in parenthesis.

Let's reword this sentence so we avoid using an ambiguous designation
based on a relative anchor (i.e. 'earlier drafts of this specification')
and just use the absolute anchor (version 1.0).

VIRTIO-164

Signed-off-by: Halil Pasic <pasic@linux.vnet.ibm.com>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>