split-ring: document in-order operation
The point of in-order for the device is to be able to skip
writing out some used descriptors, but the documentation
that we have was put in the packed ring section and
written in a way specific to the packed ring.

Writing it in a generic way is tricky, for now this patch just copies
the text from the packed ring to the split ring section, with minor
tweaks.

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/30
Suggested-by: Jason Wang <jasowang@redhat.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>