Clarifying section "2.6.12 virtqueue Operation"
We'd like to propose clarifying the usage of packets and buffers in section "2.6.12 virtqueue operation" in the note part.

This is the current writing:

"Note: As an example, the simplest virtio network device has two virtqueues: the transmit virtqueue and the receive virtqueue. The driver adds outgoing (device-readable) packets to the transmit virtqueue, and then frees them after they are used. Similarly, incoming (device-writable) buffers are added to the receive virtqueue, and processed after they are used."

And this is the proposed writing:

"Note: As an example, the simplest virtio network device has two virtqueues: the transmit virtqueue and the receive virtqueue. The driver adds outgoing packets as device-readable buffers to the transmit virtqueue, and then frees them after they are used. Similarly, device-writable buffers for incoming packets are added to the receive virtqueue, and processed after they are used."