reserve device ID for hwsim wireless simulation
The Linux mac80211-hwsim module currently allows simulation of
multiple wireless radios on a shared medium, and has an existing
API for this to work through a userspace implementation of the
medium simulation (e.g. implemented by wmediumd).

In order to simplify working with virtual machines and to enable
(time-compressed) simulation use cases, allocate a virtio device
ID to allow carrying this protocol over virtio in addition to
the current netlink sockets.

Since device ID 28 was previously requested, use 29.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Fixes: https://github.com/oasis-tcs/virtio-spec/issues/68