virtio-net: Add support for the flexible driver notification structure.
When the driver is required to send an available buffer notification
to the device, it sends the virtqueue number to be notified.

With this new feature, the device can optionally provide a per-virtqueue
value for the driver to use in driver notifications, instead of the
virtqueue number.

Some devices may benefit from this flexibility by providing, for example,
an internal virtqueue identifier, or an internal offset related to the
virtqueue number.

Changes from v8:
 * Incorporated comments for v8:
     - moved the feature from a network device to a global section
     - few minor changes

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/89
Signed-off-by: Vitaly Mireyno <vmireyno@marvell.com>
[CH: wrapped overlong lines in commit message]
Signed-off-by: Cornelia Huck <cohuck@redhat.com>