content: Minor change to clarify free_page_hint_cmd_id
The original wording was a bit unclear and could have been misinterpreted
as indicating that VIRTIO_BALLOON_FREE_PAGE_HINT was read-only instead of
the field free_page_hint_cmd_id. To clarify that break it up into two
sentences making it clear that the field is only available if the feature
is negotiated, and that the field is read-only.

Reviewed-by: Cornelia Huck <cohuck@redhat.com>
Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Acked-by: Michael S. Tsirkin <mst@redhat.com>
[CH: included under the minor cleanup rule]
Signed-off-by: Cornelia Huck <cohuck@redhat.com>