Add virtio SCMI device specification
This patch proposes a new virtio device for the Arm SCMI protocol.

The device provides a simple transport for the Arm SCMI protocol[1]. The
*S*ystem *C*ontrol and *M*anagement *I*nterface protocol allows speaking
to system controllers that allow orchestrating things like power
management, system state management and sensor access. The SCMI protocol
is used on SoCs where multiple cores and co-processors need access to
these resources.

The virtio transport allows making use of this protocol in virtualized
systems.

[1] https://developer.arm.com/docs/den0056/c

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/100
Signed-off-by: Peter Hilber <peter.hilber@opensynergy.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>