virtio-ccw: relax device requirement for revision-specific command rejection
The device is currently required to reject any command that is
not contained in the negotiated revision. Some implementations
did not actively check for the revision when processing a command;
retroactively changing these implementations can break existing
drivers.

Relaxing the rejection requirement to SHOULD makes these existing
device implementations compliant, and will not have any effect on
drivers that did not send any commands for wrong revisions.

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/96
Reviewed-by: Halil Pasic <pasic@linux.ibm.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>