Reserve more feature bits for device type usage
Feature bits 41 and above are noted as being reserved for future
extensions. However, the net device has been using bits in that space
for some time now, as it already used up the device type specific
range up to 23.

To avoid problems in the future, let's designate bits 50 to 127 to
device type specific usage (which accommodates current usage by the
net driver, and gives breathing room for future type specific bits),
and declare bits 41 to 49 and bits 128 and above to be reserved for
future extensions (which gives us some time before bit numbers move
beyond 63, which would need some changes in existing device and driver
implementations.)

Reported-by: Max Gurtovoy <mgurtovoy@nvidia.com>
Fixes: https://github.com/oasis-tcs/virtio-spec/issues/131
Reviewed-by: Max Gurtovoy <mgurtovoy@nvidia.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>