editorial: switch to upstream latexdiff
With latest latexdiff, we can finally show changed lines
within listings. Worth the switch to the lastest master.
As a bonus, we no longer need to maintain a pile of perl hacks
and complex command line as it seems to generally be smarter.
Add it as a submodule so we can track which version we use.
Notes:

1. DIFCODE_UNDERLINE file helps override colors within listings
(default is blue for addition, we use green to avoid confusion with
links elsewhere so I kept this consistent for listings).
It does not seem to be documented, something to watch out for
if/when we update latexdiff, and maybe work with the latexdiff
maintainer to document.

2. I am not sure we still need the 's/\\footnote{/\\footnote {/' hack.
Kept it around just in case. If yes it might be worth discussing
with the latexdiff maintainer to address this as part of latexdiff.

3. Invoke latexdiff as latexdiff-fast which in the past made it go faster
by using the unix diff utility. I'm not sure it's still the case but
it does not seem to hurt either.

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>