editorial: update S390 Common I/O link
Paul Knight wrote:

In running a link check, I noted several issues, none significant for this
initial public review publication:
- broken link in Normative references:
-- [S390 Common I/O] ESA/390 Common I/O-Device and Self-Description, IBM
Publication SA22-7204,
http://publibfp.dhe.ibm.com/cgi-bin/bookmgr/BOOKS/dz9ar501/CCONTENTS, and any
future revisions
(This link is currently broken, and I could not find any alternative source for
this document. - please rewrite the reference to indicate document is no longer
available at this link. You may also consider whether the solitary citation at
4.3.1.3 is still needed.)
This broken link is not a blocking issue for a Committee specification Draft,
but it will not be allowed for a Committee Specification.

Reported-by: Paul Knight <paul.knight@oasis-open.org>
Suggested-by: Halil Pasic <pasic@linux.ibm.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Reviewed-by: Cornelia Huck <cohuck@redhat.com>