virtio-net: define the VIRTIO_NET_F_CTRL_RX_EXTRA feature bit
The VIRTIO_NET_F_CTRL_RX_EXTRA feature bit is mentioned in the spec
since version 1.0, but it's not properly defined.

This patch defines the feature bit and defines the dependency on VIRTIO_NET_F_CTRL_VQ.

Since this dependency is missing in previous versions, we add it now as
a "SHOULD".

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/162

Reviewed-by: Parav Pandit <parav@nvidia.com>
Signed-off-by: Alvaro Karsz <alvaro.karsz@solid-run.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>