virtio-net: Define cfg fields before description
Currently some fields of the virtio_net_config structure are defined
before introducing the structure and some are defined after.
Better to define the configuration layout first followed by
description of all the fields.

Device configuration fields are described in the section. Change wording
from 'listed' to 'described' as suggested in patch [1].

[1] https://lists.oasis-open.org/archives/virtio-dev/202302/msg00004.html

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/161

Reviewed-by: David Edmondson <david.edmondson@oracle.com>
Signed-off-by: Parav Pandit <parav@nvidia.com>
Signed-off-by: Cornelia Huck <cohuck@redhat.com>