README.md: update mailing list info
As approved by the TC, we are moving to a less formal
way of discussing the specification, on the mailing lists
provided by the Linux Foundation:

https://groups.oasis-open.org/higherlogic/ws/public/ballot?id=3820

Update README.md, CONTRIBUTING.md and newdevice.tex accordingly.

Use this opportunity to explain when and how to use each
mailing list.

Oh yes, and device numbers are reserved through virtio-comment
not through virtio-dev. Correct that.

Message-Id: <8f5db33c96d685fcebca3579b05d09b64dd720d9.1715766697.git.mst@redhat.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>