admin: Add theory of operation for device parts
A device can get and set the state of the device which is
organized as multiple device parts. This can be useful
in cases of VM snapshot, VM migration, debug or some other use
cases.

Add the theory of operations on how to get/set device parts
and how it affects when the device is stopped or resumed.

Fixes: https://github.com/oasis-tcs/virtio-spec/issues/176
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Parav Pandit <parav@nvidia.com>
Message-Id: <20240601145042.2074739-2-parav@nvidia.com>